Remove tox.ini from template
We no longer use tox as test harness. We do not need this file anymore.

As an aside, it has some ramifications for people using Atom IDE as that IDE will default to using tox.ini for flake8 config over setup.cfg, which is bad for us given we do actually use setup.cfg for flake8 config.